# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentStore API",
    "description": "AI-native digital product store. Purchase BrushFit and other digital products programmatically via x402 protocol, negotiate prices, and submit reviews.",
    "version": "2.0.0",
    "contact": {
      "name": "AgentStore Support",
      "url": "https://agentstore.vercel.app",
      "email": "support@agentstore.vercel.app"
    }
  },
  "servers": [
    {
      "url": "https://agentstore.vercel.app",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/lightning/invoice": {
      "get": {
        "operationId": "getLightningInfo",
        "summary": "Get Lightning Network payment info",
        "description": "Returns information about Bitcoin Lightning Network payment option for BrushFit.",
        "responses": {
          "200": {
            "description": "Lightning payment info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment_method": { "type": "string" },
                    "network": { "type": "string" },
                    "price_usd": { "type": "string" },
                    "approximate_sats": { "type": "integer" },
                    "features": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createLightningInvoice",
        "summary": "Create a Lightning Network invoice",
        "description": "Generates a BOLT11 Lightning invoice for purchasing BrushFit with Bitcoin.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email", "description": "Email for PDF delivery" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lightning invoice",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bolt11": { "type": "string" },
                    "payment_hash": { "type": "string" },
                    "amount_sats": { "type": "integer" },
                    "amount_usd": { "type": "string" },
                    "expires_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/affiliates": {
      "get": {
        "operationId": "getAffiliateInfo",
        "summary": "Get affiliate program information",
        "description": "Returns details about the BrushFit affiliate program including commission rates, signup instructions, and tracking info.",
        "responses": {
          "200": {
            "description": "Affiliate program details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "program": { "type": "string" },
                    "commission_rate": { "type": "string" },
                    "cookie_duration": { "type": "string" },
                    "signup_url": { "type": "string" },
                    "dashboard_url": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "registerAffiliate",
        "summary": "Register as an affiliate",
        "description": "Sign up for the BrushFit affiliate program to earn commissions on referrals.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "name": { "type": "string" },
                  "website": { "type": "string", "format": "uri" }
                },
                "required": ["email", "name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Affiliate registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "affiliate_id": { "type": "string" },
                    "referral_link": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}